Create outline for product focused learning segment
As suggested earlier in Slack I used the outline template to get my first ideas for what a product development focused learning path segment should contain. Some of the content is based on existing patterns, some is based on in-person conversations. This is a very early very rough draft of a sketch of an outline. I do hope that based on what is there others with more experience than myself get an idea of the general direction I would like to take this and add other relevant points to it (disclaimer: I am not a product person myself, so it is very likely I miss aspects).