Sloppy translation as baseline for issue 460 (#470)
* Sloppy translation

Just a baseline for serious corrections by true french speakers.

* Committing suggested edits from reviewer

Not my translations. The content of changes are from the files posted in https://github.com/InnerSourceCommons/InnerSourceLearningPath/pull/470#issuecomment-1169218548.

I made minor formatting changes on top of the content changes.

* Update 01-introduction-fr.asciidoc

* Update 01-introduction-fr.asciidoc

This is the final version that was fully reviewed by a native French speaking IBM colleague Abdoulaye Traore. It is ready for merging into the web page.

Co-authored-by: Travis Johnson <tsjohnso@us.ibm.com>
Co-authored-by: Senthil Nathan <senthil_vela@yahoo.com>