Update 03-keeping-the-community-healthy-it.asciidoc
The automatic translation was very good, some minor problems with identifying the right gender, singular and plural when it comes to mixing Italian with English terms.
Overall, I want to mention that English and Italian have a very different way of constructing sentences. In English, sentences are usually short and simple, while in Italian we tend to do the opposite when it comes to writing documents more or less formal. For this reason, the sentences sound a bit naïve because of their simplicity, but there is not much we can do about it when it comes to translation 1:1. This works well as starting point, but I would paraphrase almost the entirety of the document. I decided to translate "comunità" back to community because it makes more sense in English than in Italian. In tech/software, we actually use the word community more. An example is open source community, were we do not translate community.